Cake Selection
Problem Statement 
You wants to buy a cake that is at least 200 grams, but you don’t want it to be heavier than 300 grams. As for sweetness, she wants her cake to have a sugar content of 50 grams or more. The cake must have 150 calories or more.

Input
The first line contain 3 integers  W, S and C the weight of the cake in grams, its sugar content in grams, and the amount of calories respectively.

Output
Print "Yes" if it is a cake that you can choose. Print "No" otherwise.

Constraints
1≤W,S,C≤500

Example 1:
Input:
200 60 160
Output:
Yes
Example 2:
Input:
370 40 340
Output:
No