Zero or One
Problem Statement
A number from 0 to 9 will be presented as an word in lower case english letter. For example- three. The program will take it as input. Print 0 if the remainder is 0 while the number is divided by 2, otherwise print 1 if the remainder is 1.

Input
The program will take a string  S as input.

Output
The output will print either 0 0r 1.

Constraints
0 ≤ |S| ≤ 1000
Example:
Enter string

Input:
three
Output:
1